Break nested types into their own files.
